add missing entities, fix typos and data types

+++ src/main/java/com/example/vibe_store/entity/Sale.java
@@ -1,9 +1,11 @@
 package com.example.vibe_store.entity;
 
+import com.example.vibe_store.entity.employee.Employee;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -16,7 +18,7 @@ public class Sale {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private Employee employee;
@@ -28,7 +30,7 @@ public class Sale {
     private Warehouse warehouse;
 
     @Column(nullable = false)
-    private double amount;
+    private BigDecimal amount;
 
     @CreationTimestamp
     private LocalDateTime salesAt;
